add categories pages
